CoPilotBot/requirements.txt
azure-ai-documentintelligence
jupyterlab
ipykernel
python-env
wandb
torch
peft
transformers
trl
pandas
huggingface_hub
bitsandbytes